Move add_file_to_bundle_if_non_existent
